chore(readme): add `with_mock` documentation
Previously the README documentation only provided an example of the
`test_with_mock` macro. No examples were provided for the `with_mock`
macro.

An example of the `with_mock` macro has now been added to the README.
Additionally information has been included to assist the user in making
their decision, on which macro is most suitable for their use case.
Guidance is given that if a given test depends upon data in the current
context, then the user should use the `with_mock` macro, as the
`test_with_mock` doesn't support a context parameter.