Implement setup_with_mocks
A helper macro that'll help avoid copy-pasting the same mock
if it is required for all unit tests in some ExUnit.Case.

For simplicity, only add support for mocking multiple modules
in setup without all the shorthand helpers when only one mock
is needed.

Wrapper macro for `setup_all` was not implemented because maintaining
the mock state across tests is difficult if one of the tests in
the module needs custom behaviour. For example, assume 9/10 tests
in some file need the same mock, but the one other test, which belongs
in the same file, does not. The one specific test can override the
'global' mock. This is easy to achieve using `setup` but more difficult
using `setup_all` and provides more developer friendly behaviour.

- Implement setup_with_mocks with support for when context is
  specified and when it is not
- Unit tests
- Move the logic of mocking multiple modules into a helper macro
- Update the README with an explanation and an example